Turn the no_std feature into default-enabled std
This also:
- fixes the `no_std` not being enabled due to a swapped `cfg_attr`.
- tests this builds with a `#![no_std]` binary.
- A lot of the cfg's have been removed due to being unneeded.
- adds a new dependency
- updates the version to 2.0
- adds a ci command to to use cargo-no-std to check no-std

Some notes on changes from #68:

Everything in `core` is a subset of things in `std`.
`core` is available in `std` environments, so if you're building
a `no_std`-compatible library, it's best to import things that
don't _require_ `std` from `core.

core2 is also meant to be used as a `no_std` polyfill with its
`std` feature. That is, you can reference `core2::io`, and if the
`std` feature is enabled, it actually references `std::io`.